backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
requests==2.32.3
transformers==4.44.2
torch==2.3.1
sentencepiece==0.2.0
sentence-transformers==2.2.2
spacy==3.7.5
nltk==3.9.1
networkx==3.3
plotly==5.24.0
pandas==2.2.2
numpy==1.26.4
beautifulsoup4==4.12.3
feedparser==6.0.11
wordcloud==1.9.3
